Test creation and file movement Moved MersenneTwister and Shuffle to the Model section Modified DeckTest and TableTest
