+++ Assignment1/src/main/java/model/MersenneTwister.java
@@ -1,4 +1,4 @@
-package controllers;
+package model;
 
 import java.io.*;
