+++ Assignment1/src/test/java/model/TableTest.java
@@ -1,16 +1,14 @@
 package model;
 
 import org.junit.Test;
-
-import static org.junit.Assert.assertNotNull;
-import static org.junit.Assert.assertThat;
+import static org.junit.Assert.*;
 
 /**
  * Created by Brandon on 1/21/2016.
  */
 public class TableTest {
     @Test
-    public void testArrayListSize() {
+    public void testArrayListContents() {
         TableCreator table = new TableCreator();
         assertNotNull(table);
     }
